cleaned code and updated docker dependencies
